// types/StudentTasksAnswersFormState.ts
export type StudentTasksAnswersFormState = {
  didLoaded: boolean;
  didPdfLoaded: boolean;
  isExpertMode: boolean;
  isVisibleTranslate: boolean;
  textIsVisible: boolean,
  exercisesIsVisible: boolean;
  isVisibleUploadAnswerComponent: boolean;
  showWithoutNikkudot: boolean;
  isSingleGameTask: boolean;
};